requirements.txt
python-dotenv>=1.0.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.2
langchain-huggingface>=0.0.2
langchain-community>=0.0.14
uvicorn>=0.23.2
fastapi>=0.104.0
langserve[all]>=0.0.34
langsmith>=0.0.83
pinecone-client>=3.0.0
langchain-pinecone>=0.0.4
unstructured>=0.10.30
unstructured[pdf]>=0.10.30
mistralai>=0.0.11
langchain-mistralai>=0.0.3
huggingface-hub>=0.17.3
tqdm>=4.66.1